Fix observer command & button
